(#21) feat: 한 파일로 묶기
